requirements.txt
flask
pillow
mediapipe
numpy
torch
torchvision
torchaudio
google-generativeai
diffusers
transformers
accelerate
safetensors
opencv-python
gunicorn